Updated README.md to Reflect changes/additions to script
README now reflects the menu based operation of the script in addtion to the changes made to gitignore functions and node_modules function operations. Removed redundant chmod step from installation instructions as the recommendation to setup script alias covers that intent of that command. Fixed headers and steps in Installation/Update section to match removals/additions